# backend/docs/openapi.yaml
openapi: 3.0.3
info:
  title: SocialPredict API
  version: 0.1.0
  description: >
    HTTP contract for the SocialPredict backend service. This spec is generated
    from the Go handlers and DTOs under backend/server and backend/handlers.
    Routes wrapped by the shared security middleware can return a shared
    transport-level `429 text/plain` response before reaching handler-specific
    JSON envelope logic when the rate limit is exceeded. Per-operation `429`
    responses document that middleware behavior where it applies.
servers:
  - url: http://localhost:8080
    description: Local dev server
...
paths:
...
  /v0/markets:
    get:
      operationId: listMarkets
      tags: [Markets]
      summary: List markets
      description: >
        Returns markets with optional filtering by status and creator. When status is
        provided, results are backed by the status-specific listing path used by the
        live handler.
      parameters:
        - in: query
          name: status
          required: false
          description: >
            Filter by market status. Accepted values are active, closed, resolved, and all
            (or omitted for all markets).
          schema:
            type: string
            enum: [active, closed, resolved, all]
        - in: query
          name: limit
          required: false
          description: Maximum number of markets to return. Malformed, non-positive, or out-of-range values fall back to the live default of 50.
          schema:
            type: integer
        - in: query
          name: offset
          required: false
          description: Number of markets to skip before collecting results. Malformed or negative values fall back to 0.
          schema:
            type: integer
        - in: query
          name: created_by
          required: false
          description: Only include markets created by the supplied username.
          schema:
            type: string
      responses:
        '200':
          description: Markets returned successfully.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ListMarketsResponse'
        '400':
          description: Invalid or unsupported status filter.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ReasonResponse'
              examples:
                invalidRequest:
                  value:
                    ok: false
                    reason: INVALID_REQUEST
        '500':
          description: Unexpected server error while listing markets.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ReasonResponse'
              examples:
                internalError:
                  value:
                    ok: false
                    reason: INTERNAL_ERROR
        '429':
          $ref: '#/components/responses/MiddlewareRateLimitExceeded'
    post:
      operationId: createMarket
      tags: [Markets]
      summary: Create a market
      description: Creates a new prediction market for the authenticated user.
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateMarketRequest'
      responses:
        '201':
          description: Market created successfully.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MarketResponse'
        '400':
          description: Request decoding failed, validation failed, or the creator lacks funds.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ReasonResponse'
              examples:
                invalidRequest:
                  value:
                    ok: false
                    reason: INVALID_REQUEST
                validationFailed:
                  value:
                    ok: false
                    reason: VALIDATION_FAILED
                insufficientBalance:
                  value:
                    ok: false
                    reason: INSUFFICIENT_BALANCE
        '401':
          description: Authentication is required.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ReasonResponse'
              examples:
                authenticationRequired:
                  value:
                    ok: false
                    reason: AUTHENTICATION_REQUIRED
        '403':
          description: Password change required before creating markets.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ReasonResponse'
              examples:
                passwordChangeRequired:
                  value:
                    ok: false
                    reason: PASSWORD_CHANGE_REQUIRED
        '404':
          description: Authenticated user record could not be loaded.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ReasonResponse'
              examples:
                notFound:
                  value:
                    ok: false
                    reason: NOT_FOUND
        '500':
          description: Unexpected server error during market creation.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ReasonResponse'
              examples:
                internalError:
                  value:
                    ok: false
                    reason: INTERNAL_ERROR
        '429':
          $ref: '#/components/responses/MiddlewareRateLimitExceeded'
  /v0/markets/search:
    get:
      operationId: searchMarkets
      tags: [Markets]
      summary: Search markets
      description: >
        Searches markets by question title with optional status filtering. Supports both
        "query" and legacy "q" parameters.
      parameters:
        - in: query
          name: query
          required: false
          description: Search query (preferred parameter name).
          schema:
            type: string
        - in: query
          name: q
          required: false
          description: Legacy search query parameter (used if "query" is not provided).
          schema:
            type: string
        - in: query
          name: status
          required: false
          description: Filter by market status; same accepted values as /v0/markets.
          schema:
            type: string
            enum: [active, closed, resolved, all]
        - in: query
          name: limit
          required: false
          description: Maximum number of primary results to return. Malformed, non-positive, or out-of-range values fall back to the live default of 20.
          schema:
            type: integer
        - in: query
          name: offset
          required: false
          description: Number of primary results to skip before collecting. Malformed or negative values fall back to 0.
          schema:
            type: integer
      responses:
        '200':
          description: Search completed successfully.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SearchResponse'
        '400':
          description: Missing query parameter or invalid status.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ReasonResponse'
              examples:
                invalidRequest:
                  value:
                    ok: false
                    reason: INVALID_REQUEST
        '500':
          description: Unexpected server error during search.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ReasonResponse'
              examples:
                internalError:
                  value:
                    ok: false
                    reason: INTERNAL_ERROR
        '429':
          $ref: '#/components/responses/MiddlewareRateLimitExceeded'
  /v0/markets/status:
    get:
      operationId: listAllMarketStatuses
      tags: [Markets]
      summary: List markets across all statuses
      description: Convenience endpoint equivalent to `/v0/markets?status=all`.
      parameters:
        - in: query
          name: limit
          required: false
          description: Maximum number of markets to return. Malformed, non-positive, or out-of-range values fall back to the live default of 100.
          schema:
            type: integer
        - in: query
          name: offset
          required: false
          description: Number of markets to skip before collecting results. Malformed or negative values fall back to 0.
          schema:
            type: integer
      responses:
        '200':
          description: Markets returned successfully.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ListMarketsResponse'
        '500':
          description: Unexpected server error while listing markets.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ReasonResponse'
              examples:
                internalError:
                  value:
                    ok: false
                    reason: INTERNAL_ERROR
        '429':
          $ref: '#/components/responses/MiddlewareRateLimitExceeded'
...
  /v0/markets/{id}:
    get:
      operationId: getMarketDetails
      tags: [Markets]
      summary: Get market details
      description: >
        Fetches a single market by its identifier, including probability history and
        aggregate statistics.
      parameters:
        - in: path
          name: id
          required: true
          description: Numeric identifier of the market.
          schema:
            type: integer
            format: int64
      responses:
        '200':
          description: Market returned successfully.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MarketDetailsResponse'
        '400':
          description: Invalid market ID.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ReasonResponse'
              examples:
                invalidRequest:
                  value:
                    ok: false
                    reason: INVALID_REQUEST
        '404':
          description: Market not found.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ReasonResponse'
              examples:
                notFound:
                  value:
                    ok: false
                    reason: NOT_FOUND
        '500':
          description: Unexpected server error while fetching market details.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ReasonResponse'
              examples:
                internalError:
                  value:
                    ok: false
                    reason: INTERNAL_ERROR
        '429':
          $ref: '#/components/responses/MiddlewareRateLimitExceeded'
  /v0/markets/{id}/resolve:
    post:
      operationId: resolveMarket
      tags: [Markets]
      summary: Resolve a market
      description: Sets the final outcome for a market once it has concluded.
      security:
        - bearerAuth: []
      parameters:
        - in: path
          name: id
          required: true
          description: Numeric identifier of the market to resolve.
          schema:
            type: integer
            format: int64
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ResolveMarketRequest'
      responses:
        '204':
          description: Market resolved successfully; no content is returned.
        '400':
          description: Invalid request body or validation failure in the resolution payload.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ReasonResponse'
              examples:
                invalidRequest:
                  value:
                    ok: false
                    reason: INVALID_REQUEST
                validationFailed:
                  value:
                    ok: false
                    reason: VALIDATION_FAILED
        '401':
          description: Authentication is required.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ReasonResponse'
              examples:
                authenticationRequired:
                  value:
                    ok: false
                    reason: AUTHENTICATION_REQUIRED
        '403':
          description: Password change requirements or creator authorization block resolution.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ReasonResponse'
              examples:
                passwordChangeRequired:
                  value:
                    ok: false
                    reason: PASSWORD_CHANGE_REQUIRED
                authorizationDenied:
                  value:
                    ok: false
                    reason: AUTHORIZATION_DENIED
        '404':
          description: Market or authenticated user record not found.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ReasonResponse'
              examples:
                notFound:
                  value:
                    ok: false
                    reason: NOT_FOUND
        '409':
          description: Market is already resolved or otherwise in an invalid state for resolution.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ReasonResponse'
              examples:
                validationFailed:
                  value:
                    ok: false
                    reason: VALIDATION_FAILED
        '500':
          description: Unexpected server error during resolution.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ReasonResponse'
              examples:
                internalError:
                  value:
                    ok: false
                    reason: INTERNAL_ERROR
        '429':
          $ref: '#/components/responses/MiddlewareRateLimitExceeded'
  /v0/markets/{id}/leaderboard:
    get:
      operationId: getMarketLeaderboard
      tags: [Markets]
      summary: Get market leaderboard
      description: Ranks participants in a single market by profit and traded volume.
      parameters:
        - in: path
          name: id
          required: true
          description: Numeric identifier of the market.
          schema:
            type: integer
            format: int64
        - in: query
          name: limit
          required: false
          description: Maximum number of rows to return. Malformed, non-positive, or out-of-range values fall back to the live default of 100.
          schema:
            type: integer
        - in: query
          name: offset
          required: false
          description: Number of rows to skip before collecting results. Malformed or negative values fall back to 0.
          schema:
            type: integer
      responses:
        '200':
          description: Leaderboard returned successfully.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MarketLeaderboardResponse'
        '400':
          description: Invalid market ID or pagination parameters.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ReasonResponse'
              examples:
                invalidRequest:
                  value:
                    ok: false
                    reason: INVALID_REQUEST
        '404':
          description: Market not found.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ReasonResponse'
              examples:
                notFound:
                  value:
                    ok: false
                    reason: NOT_FOUND
        '500':
          description: Unexpected server error while computing leaderboard.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ReasonResponse'
              examples:
                internalError:
                  value:
                    ok: false
                    reason: INTERNAL_ERROR
        '429':
          $ref: '#/components/responses/MiddlewareRateLimitExceeded'
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  responses:
    MiddlewareRateLimitExceeded:
      description: Request rejected by shared rate-limiting middleware before handler-specific logic ran.
      content:
        text/plain:
          schema:
            $ref: '#/components/schemas/PlainTextErrorResponse'
  schemas:
...
    ReasonResponse:
      type: object
      properties:
        ok:
          type: boolean
          enum: [false]
        reason:
          $ref: '#/components/schemas/FailureReason'
      required: [ok, reason]
    PlainTextErrorResponse:
      type: string
      description: Plain text error emitted by shared middleware before request-specific handlers run.
...
    CreateMarketRequest:
      type: object
      required: [questionTitle, outcomeType, resolutionDateTime]
      properties:
        questionTitle:
          type: string
          maxLength: 160
        description:
          type: string
          maxLength: 2000
        outcomeType:
          type: string
        resolutionDateTime:
          type: string
          format: date-time
        yesLabel:
          type: string
          maxLength: 20
        noLabel:
          type: string
          maxLength: 20
...
    MarketResponse:
      type: object
      properties:
        id:
          type: integer
          format: int64
        questionTitle:
          type: string
        description:
          type: string
        outcomeType:
          type: string
        resolutionDateTime:
          type: string
          format: date-time
        creatorUsername:
          type: string
        yesLabel:
          type: string
        noLabel:
          type: string
        status:
          type: string
        isResolved:
          type: boolean
        resolutionResult:
          type: string
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
...
    ListMarketsResponse:
      type: object
      properties:
        markets:
          type: array
          items:
            $ref: '#/components/schemas/MarketOverviewResponse'
        total:
          type: integer
    MarketDetailsResponse:
      type: object
      properties:
        market:
          $ref: '#/components/schemas/PublicMarketResponse'
        creator:
          $ref: '#/components/schemas/CreatorResponse'
        probabilityChanges:
          type: array
          items:
            $ref: '#/components/schemas/ProbabilityChange'
        numUsers:
          type: integer
        totalVolume:
          type: integer
          format: int64
        marketDust:
          type: integer
          format: int64
...
    SearchResponse:
      type: object
      properties:
        primaryResults:
          type: array
          items:
            $ref: '#/components/schemas/MarketOverviewResponse'
        fallbackResults:
          type: array
          items:
            $ref: '#/components/schemas/MarketOverviewResponse'
        query:
          type: string
        primaryStatus:
          type: string
        primaryCount:
          type: integer
        fallbackCount:
          type: integer
        totalCount:
          type: integer
        fallbackUsed:
          type: boolean
    ResolveMarketRequest:
      type: object
      required: [resolution]
      properties:
        resolution:
          type: string
          description: Case-insensitive resolution outcome accepted by the live handler and domain service (`YES`, `NO`, or `N/A`).
    MarketLeaderboardResponse:
      type: object
      properties:
        marketId:
          type: integer
          format: int64
        leaderboard:
          type: array
          items:
            $ref: '#/components/schemas/LeaderboardRow'
        total:
          type: integer
...